services/proxy/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
PyJWT==2.9.0
pydantic-settings==2.5.2
cryptography==43.0.3
